requirements.txt
fastapi[all]
uvicorn
pydantic
psycopg
pytest-testdox
python-dotenv